requirements.txt
selenium>=4.16.0
webdriver-manager>=4.0.1
pytest>=7.4.3
pytest-html>=4.1.1
pytest-xdist>=3.5.0
faker>=22.0.0
python-dotenv>=1.0.0
colorlog>=6.8.0
Pillow>=10.1.0
